mv: Move get_sync_msgs function to fim_synchronization
